add excel

+++ back-end/src/services/file.service.ts
@@ -2,7 +2,7 @@
  * @Author: sunji 2025506282@qq.com
  * @Date: 2022-09-16 13:34:34
  * @LastEditors: sunji 2025506282@qq.com
- * @LastEditTime: 2022-09-19 15:25:42
+ * @LastEditTime: 2022-09-20 15:33:47
  * @FilePath: \back-end\src\services\file.service.ts
  * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
  */
@@ -31,25 +31,40 @@ export class FileService {
     try {
       const filePath = path.resolve(UPLOAD_FILES_DIR, fileHash + `.${suffix}`);
       const chunksDir = path.resolve(UPLOAD_FILES_DIR, fileHash);
-      console.log("指定位置创建可写流", chunksDir, filePath);
       const chunkPaths = await fse.readdir(chunksDir);
       chunkPaths.sort(
         (a: string, b: string) =>
-          Number(a.split("_")[1]) - Number(b.split("_")[1])
+          Number(a.split("___$")[1]) - Number(b.split("___$")[1])
       );
-      await Promise.all(
-        chunkPaths.map((chunkPath: string, index: number) =>
-          pipeStream(
-            path.resolve(chunksDir, chunkPath),
-            // 指定位置创建可写流
-            fse.createWriteStream(filePath, {
-              start: index * size,
-              // end: (index + 1) * size,
-            })
+      if (!fse.existsSync(filePath)) {
+        console.log(
+          " --------------start------filePath----------------------",
+          chunksDir,
+          filePath
+        );
+        Promise.all(
+          chunkPaths.map((chunkPath: string, index: number) =>
+            pipeStream(
+              path.resolve(chunksDir, chunkPath),
+              // 指定位置创建可写流
+              fse.createWriteStream(filePath, {
+                start: index * size,
+                // end: (index + 1) * size,
+              })
+            )
           )
         )
-      );
-      fse.rmdirSync(chunksDir);
+          .then((res) => {
+            console.log("rmdirSync-----------:", chunksDir);
+            fse.rmdirSync(chunksDir);
+          })
+          .catch((err) => {
+            console.log("err----------------:", err);
+          })
+          .finally(() => {
+            console.log("finally-------");
+          });
+      }
       return true;
     } catch (err) {
       console.log("err:", err);
@@ -70,12 +85,11 @@ export class FileService {
         const [chunk] = files.file;
         const [hash] = fields.hash;
         // 注意这里的hash包含文件的hash和块的索引，所以需要使用split切分
-        const chunksDir = path.resolve(UPLOAD_FILES_DIR, hash.split("_")[0]);
+        const chunksDir = path.resolve(UPLOAD_FILES_DIR, hash.split("___$")[0]);
         if (!fse.existsSync(chunksDir)) {
           await fse.mkdirs(chunksDir);
         }
         if (!fse.existsSync(chunksDir + "/" + hash)) {
-          console.log("files: ", files, fields, chunksDir);
           await fse.move(chunk.path, chunksDir + "/" + hash, {
             overwrite: true,
           });
